ROC-6393: Fix conflicts

+++ src/integration-test/tests/citizen/claim/enterDetails_test.ts
@@ -16,7 +16,9 @@ Feature('Claimant Enter details of claim')
 
 Scenario('I can prepare a claim with no interest @citizen', { retries: 3 }, async (I: I) => {
   userSteps.login(userSteps.getClaimantEmail())
-  testingSupport.deleteClaimDraft()
+  if (process.env.FEATURE_TESTING_SUPPORT === 'true') {
+    testingSupport.deleteClaimDraft()
+  }
   claimSteps.completeEligibility()
   claimSteps.completeStartOfClaimJourney(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL, true)
   interestSteps.skipClaimInterest()
@@ -54,7 +56,9 @@ Scenario('I can prepare a claim with no interest @citizen', { retries: 3 }, asyn
 
 Scenario('I can prepare a claim with different interest rate and date @citizen', { retries: 3 }, async (I: I) => {
   userSteps.login(userSteps.getClaimantEmail())
-  testingSupport.deleteClaimDraft()
+  if (process.env.FEATURE_TESTING_SUPPORT === 'true') {
+    testingSupport.deleteClaimDraft()
+  }
   claimSteps.completeEligibility()
   claimSteps.completeStartOfClaimJourney(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL, true)
   interestSteps.enterSpecificInterestRateAndDate(2, '1990-01-01')
@@ -64,12 +68,16 @@ Scenario('I can prepare a claim with different interest rate and date @citizen',
   claimSteps.enterClaimDetails()
   userSteps.selectCheckAndSubmitYourClaim()
   I.see('£80.50')
-  testingSupport.deleteClaimDraft()
+  if (process.env.FEATURE_TESTING_SUPPORT === 'true') {
+    testingSupport.deleteClaimDraft()
+  }
 })
 
 Scenario('I can prepare a claim with a manually entered interest amount and a daily amount added @citizen', { retries: 3 }, async (I: I) => {
   userSteps.login(userSteps.getClaimantEmail())
-  testingSupport.deleteClaimDraft()
+  if (process.env.FEATURE_TESTING_SUPPORT === 'true') {
+    testingSupport.deleteClaimDraft()
+  }
   claimSteps.completeEligibility()
   claimSteps.completeStartOfClaimJourney(PartyType.INDIVIDUAL, PartyType.INDIVIDUAL, true)
   interestSteps.enterBreakdownInterestAmountAndDailyAmount()
@@ -81,7 +89,9 @@ Scenario('I can prepare a claim with a manually entered interest amount and a da
   I.see('£80.50')
   I.see('Break down interest for different time periods or items')
   I.see('Show how you calculated the amount')
-  testingSupport.deleteClaimDraft()
+  if (process.env.FEATURE_TESTING_SUPPORT === 'true') {
+    testingSupport.deleteClaimDraft()
+  }
 })
 
 // The @citizen-smoke-test tag used for running smoke tests with pre-registered user
